+++ epstein_files/documents/messenger_log.py
@@ -13,6 +13,7 @@ from epstein_files.documents.documents.doc_cfg import Metadata
 from epstein_files.documents.imessage.text_message import TextMessage
 from epstein_files.output.highlight_config import styled_name
 from epstein_files.output.rich import LAST_TIMESTAMP_STYLE, build_table, highlighter
+from epstein_files.people.interesting_people import TEXTERS_OF_INTEREST
 from epstein_files.util.constant.names import JEFFREY_EPSTEIN, Name
 from epstein_files.util.constant.strings import AUTHOR, TIMESTAMP_STYLE
 from epstein_files.util.helpers.data_helpers import days_between, days_between_str, iso_timestamp, sort_dict
@@ -35,6 +36,14 @@ class MessengerLog(Communication):
     @property
     def border_style(self) -> str:
         return self.author_style
+
+    @property
+    def is_interesting(self) -> bool | None:
+        """Junk emails are not interesting."""
+        if (is_interesting := super().is_interesting) is not None:
+            return is_interesting
+        elif self.author is None or self.author in TEXTERS_OF_INTEREST:
+            return True
 
     @property
     def metadata(self) -> Metadata:
@@ -66,6 +75,7 @@ class MessengerLog(Communication):
     def __post_init__(self):
         super().__post_init__()
         self.messages = [self._build_message(match) for match in MSG_REGEX.finditer(self.text)]
+        self.recipients = [JEFFREY_EPSTEIN]  # Epstein is always the counterparty for these files
 
     def first_message_at(self, name: Name) -> datetime:
         return self.messages_by(name)[0].parse_timestamp()