+++ epstein_files/output/output.py
@@ -14,6 +14,7 @@ from epstein_files.epstein_files import EpsteinFiles, count_by_month
 from epstein_files.output.rich import *
 from epstein_files.output.title_page import (print_color_key, print_other_page_link, print_section_header,
      print_section_summary_table)
+from epstein_files.people.interesting_people import EMAILERS_TO_PRINT
 from epstein_files.people.person import Person
 from epstein_files.util.constant.html import CONSOLE_HTML_FORMAT, HTML_TERMINAL_THEME, PAGE_TITLE
 from epstein_files.util.constant.names import *
@@ -30,39 +31,26 @@ DEVICE_SIGNATURE_PADDING = (1, 0)
 OTHER_INTERESTING_EMAILS_SUBTITLE = 'Other Interesting Emails\n(these emails have been flagged as being of particular interest)'
 PRINT_COLOR_KEY_EVERY_N_EMAILS = 150
 
-# Order matters. Default names to print emails for.
-DEFAULT_EMAILERS = [
-    BROCK_PIERCE,
-    JABOR_Y,
-    JOI_ITO,
-    STEVEN_SINOFSKY,
-    AMIR_TAAKI,
-    DANIEL_SIAD,
-    AL_SECKEL,
-    AUSTIN_HILL,
-    VINCENZO_IOZZO,
-    MARIA_PRUSAKOVA,
-    JEREMY_RUBIN,
-    JEAN_LUC_BRUNEL,
-    DAVID_STERN,
-    STEVEN_HOFFENBERG,
-    EHUD_BARAK,
-    MASHA_DROKOVA,
-    STEVE_BANNON,
-    JULIA_SANTOS,
-    TYLER_SHEARS,
-    SERGEY_BELYAKOV,
-    GANBAT_CHULUUNKHUU,
-    RENATA_BOLOTOVA,
-    CHRISTINA_GALBRAITH,
-    MOHAMED_WAHEED_HASSAN,
-    JENNIFER_JACQUET,
-    ZUBAIR_KHAN,
-    ROSS_GOW,
-    DAVID_BLAINE,
-    None,
-    JEFFREY_EPSTEIN,
-]
+
+def print_curated_chronological(epstein_files: EpsteinFiles) -> list[Document]:
+    other_files_queue = []  # Collect other files into tables
+    printed_docs: list[Document] = []
+
+    for doc in epstein_files.unique_documents:
+        if not doc.is_interesting:
+            continue
+        elif isinstance(doc, OtherFile):
+            other_files_queue.append(doc)
+            continue
+        elif other_files_queue:
+            console.print(Padding(OtherFile.files_preview_table(other_files_queue), (0, 0, 1, 2)))
+            printed_docs.extend(other_files_queue)
+            other_files_queue = []
+
+        console.print(doc)
+        printed_docs.append(doc)
+
+    return printed_docs
 
 
 def print_doj_files(epstein_files: EpsteinFiles) -> list[DojFile | Email]:
@@ -150,7 +138,7 @@ def print_emails_section(epstein_files: EpsteinFiles) -> list[Email]:
         if args.all_emails:
             people_to_print = all_emailers
         else:
-            people_to_print = epstein_files.person_objs(DEFAULT_EMAILERS)
+            people_to_print = epstein_files.person_objs(EMAILERS_TO_PRINT)
 
         print_section_summary_table(Person.emailer_info_table(all_emailers, people_to_print))